requirements/test-requirements.txt
# a mirror of ../pyproject.toml
# used for caching dependencies in github action
chex >= 0.1.0
flax >= 0.5.0
pytest >= 6.0.0
gymnasium >= 0.29.0
ray >= 2.0.0
tensorflow-datasets >= 4.0.0,
grain >= 0.1.0,
envpool >= 0.8.0
gpjax >= 0.8.0
optax >= 0.1.7